fix SubscribersNotificationTest, broadcast_change to Helpers
